Update README, add NEWLIB.md

+++ emulator/build.sh
@@ -7,6 +7,10 @@ cmake .. -DCMAKE_BUILD_TYPE=Release
 make -j6
 popd
 
-ln -fs .build/rvmicro .
-ln -fs .build/rvnewlib .
+if test -f ".build/rvmicro"; then
+	ln -fs .build/rvmicro .
+fi
+if test -f ".build/rvnewlib"; then
+	ln -fs .build/rvnewlib .
+fi
 ln -fs .build/rvlinux .
